+++ frontend/src/AppRoutes/index.tsx
@@ -4,6 +4,7 @@ import Spinner from 'components/Spinner';
 import AppLayout from 'container/AppLayout';
 import { useThemeConfig } from 'hooks/useDarkMode';
 import { NotificationProvider } from 'hooks/useNotifications';
+import { ResourceProvider } from 'hooks/useResourceAttribute';
 import history from 'lib/history';
 import { QueryBuilderProvider } from 'providers/QueryBuilder';
 import React, { Suspense } from 'react';
@@ -17,30 +18,32 @@ function App(): JSX.Element {
 
 	return (
 		<ConfigProvider theme={themeConfig}>
-			<NotificationProvider>
-				<Router history={history}>
+			<Router history={history}>
+				<NotificationProvider>
 					<PrivateRoute>
-						<QueryBuilderProvider>
-							<AppLayout>
-								<Suspense fallback={<Spinner size="large" tip="Loading..." />}>
-									<Switch>
-										{routes.map(({ path, component, exact }) => (
-											<Route
-												key={`${path}`}
-												exact={exact}
-												path={path}
-												component={component}
-											/>
-										))}
+						<ResourceProvider>
+							<QueryBuilderProvider>
+								<AppLayout>
+									<Suspense fallback={<Spinner size="large" tip="Loading..." />}>
+										<Switch>
+											{routes.map(({ path, component, exact }) => (
+												<Route
+													key={`${path}`}
+													exact={exact}
+													path={path}
+													component={component}
+												/>
+											))}
 
-										<Route path="*" component={NotFound} />
-									</Switch>
-								</Suspense>
-							</AppLayout>
-						</QueryBuilderProvider>
+											<Route path="*" component={NotFound} />
+										</Switch>
+									</Suspense>
+								</AppLayout>
+							</QueryBuilderProvider>
+						</ResourceProvider>
 					</PrivateRoute>
-				</Router>
-			</NotificationProvider>
+				</NotificationProvider>
+			</Router>
 		</ConfigProvider>
 	);
 }
