--- frontend/src/container/MetricsApplication/ResourceAttributesFilter/types.ts
@@ -1,11 +1,0 @@
-export interface IOption {
-	label: string;
-	value: string;
-}
-
-export interface IResourceAttributeQuery {
-	id: string;
-	tagKey: string;
-	operator: string;
-	tagValue: string[];
-}
