+++ frontend/src/container/ResourceAttributesFilter/components/QueryChip/index.ts
@@ -1,0 +1,3 @@
+import QueryChip from './QueryChip';
+
+export default QueryChip;
