+++ frontend/src/container/ResourceAttributesFilter/components/QueryChip/types.ts
@@ -1,0 +1,6 @@
+import { IResourceAttribute } from 'hooks/useResourceAttribute/types';
+
+export interface IQueryChipProps {
+	queryData: IResourceAttribute;
+	onClose: (id: string) => void;
+}
