+++ frontend/src/container/ResourceAttributesFilter/index.ts
@@ -1,0 +1,3 @@
+import ResourceAttributesFilter from './ResourceAttributesFilter';
+
+export default ResourceAttributesFilter;
