+++ frontend/src/hooks/useResourceAttribute/context.ts
@@ -1,0 +1,7 @@
+import { createContext } from 'react';
+
+import { IResourceAttributeProps } from './types';
+
+export const ResourceContext = createContext<IResourceAttributeProps>(
+	{} as IResourceAttributeProps,
+);
