+++ frontend/src/hooks/useResourceAttribute/index.ts
@@ -1,0 +1,7 @@
+import ResourceProvider from './ResourceProvider';
+import useResourceAttribute from './useResourceAttribute';
+import { convertMetricKeyToTrace, isResourceEmpty } from './utils';
+
+export default useResourceAttribute;
+
+export { convertMetricKeyToTrace, isResourceEmpty, ResourceProvider };
