+++ frontend/src/pages/AllErrors/index.tsx
@@ -1,6 +1,7 @@
 import RouteTab from 'components/RouteTab';
 import ROUTES from 'constants/routes';
 import AllErrorsContainer from 'container/AllError';
+import ResourceAttributesFilter from 'container/ResourceAttributesFilter';
 import React from 'react';
 import { useTranslation } from 'react-i18next';
 
@@ -8,18 +9,21 @@ function AllErrors(): JSX.Element {
 	const { t } = useTranslation();
 
 	return (
-		<RouteTab
-			{...{
-				routes: [
-					{
-						Component: AllErrorsContainer,
-						name: t('routes.all_errors'),
-						route: ROUTES.ALL_ERROR,
-					},
-				],
-				activeKey: t('routes.all_errors'),
-			}}
-		/>
+		<>
+			<ResourceAttributesFilter />
+			<RouteTab
+				{...{
+					routes: [
+						{
+							Component: AllErrorsContainer,
+							name: t('routes.all_errors'),
+							route: ROUTES.ALL_ERROR,
+						},
+					],
+					activeKey: t('routes.all_errors'),
+				}}
+			/>
+		</>
 	);
 }
