+++ frontend/src/pages/Metrics/index.tsx
@@ -3,10 +3,11 @@ import getLocalStorageKey from 'api/browser/localstorage/get';
 import ReleaseNote from 'components/ReleaseNote';
 import Spinner from 'components/Spinner';
 import { SKIP_ONBOARDING } from 'constants/onboarding';
-import ResourceAttributesFilter from 'container/MetricsApplication/ResourceAttributesFilter';
 import MetricTable from 'container/MetricsTable';
+import ResourceAttributesFilter from 'container/ResourceAttributesFilter';
 import { useNotifications } from 'hooks/useNotifications';
-import { convertRawQueriesToTraceSelectedTags } from 'lib/resourceAttributes';
+import useResourceAttribute from 'hooks/useResourceAttribute';
+import { convertRawQueriesToTraceSelectedTags } from 'hooks/useResourceAttribute/utils';
 import React, { useEffect, useMemo } from 'react';
 import { connect, useSelector } from 'react-redux';
 import { useLocation } from 'react-router-dom';
@@ -25,12 +26,9 @@ function Metrics({ getService }: MetricsProps): JSX.Element {
 		GlobalReducer
 	>((state) => state.globalTime);
 	const location = useLocation();
-	const {
-		services,
-		resourceAttributeQueries,
-		error,
-		errorMessage,
-	} = useSelector<AppState, MetricReducer>((state) => state.metrics);
+	const { services, error, errorMessage } = useSelector<AppState, MetricReducer>(
+		(state) => state.metrics,
+	);
 	const { notifications } = useNotifications();
 
 	useEffect(() => {
@@ -41,12 +39,13 @@ function Metrics({ getService }: MetricsProps): JSX.Element {
 		}
 	}, [error, errorMessage, notifications]);
 
+	const { queries } = useResourceAttribute();
+
 	const selectedTags = useMemo(
-		() =>
-			(convertRawQueriesToTraceSelectedTags(resourceAttributeQueries) as Tags[]) ||
-			[],
-		[resourceAttributeQueries],
+		() => (convertRawQueriesToTraceSelectedTags(queries, '') as Tags[]) || [],
+		[queries],
 	);
+
 	const isSkipped = getLocalStorageKey(SKIP_ONBOARDING) === 'true';
 
 	useEffect(() => {
