+++ frontend/src/types/reducer/metrics.ts
@@ -1,4 +1,3 @@
-import { IResourceAttributeQuery } from 'container/MetricsApplication/ResourceAttributesFilter/types';
 import { DBOverView } from 'types/api/metrics/getDBOverview';
 import { ExternalAverageDuration } from 'types/api/metrics/getExternalAverageDuration';
 import { ExternalError } from 'types/api/metrics/getExternalError';
@@ -19,8 +18,6 @@ interface MetricReducer {
 	externalAverageDuration: ExternalAverageDuration[];
 	externalError: ExternalError[];
 	serviceOverview: ServiceOverview[];
-	resourceAttributeQueries: IResourceAttributeQuery[];
-	resourceAttributePromQLQuery: string;
 	topLevelOperations: string[];
 }
